+++ utils/utils.py
@@ -35,7 +35,9 @@ class Utils:
 
     @staticmethod
     def remove_file(file):
-        os.remove(file)
+        print('Removing: ',file)
+        if Utils.exists(file):
+            os.remove(file)
 
     @staticmethod
     def remove_dir(folder):
@@ -113,7 +115,7 @@ class Utils:
         :param path: LIST of args
         :return:
         '''
-        print(args)
+        print(' '.join(args))
         if out is None:
             subprocess.call(args)
         else:
@@ -204,11 +206,16 @@ class StatsReport:
         return dist[rows-1][cols-1]
 
 from Bio import SearchIO, SeqIO
+from Bio.Seq import Seq
 from Bio.File import as_handle
+from Bio.Alphabet import generic_dna
 from Bio.SearchIO._model import QueryResult, Hit, HSP, HSPFragment
 from Bio.SearchIO._utils import get_processor
 
 class BioUtils:
+    @staticmethod
+    def reverse_complement(chain):
+        return str(Seq(chain, generic_dna).reverse_complement())
     @staticmethod
     def identicalClustering(file, program = 'mmseqs', args = ['tmp/clusters','tmp/', '-c','1.0','--min-seq-id',
                                                               '1.0','-v','3','--seq-id-mode','0','--cov-mode','1','--remove-tmp-files'
